Create ssh_repo_elm.md
I tried to build my first reproducible analysis with datalad repos for input / output, using data hosted on elm via ssh. And I struggled :) so I put together this tutorial with the help of chatgpt. I think I got it to work, but it's very possible I made mistakes. And I really struggled with datalad's docs. So hopefully more knowledgeable people can review and confirm this is in order, and the tutorial can save time to others (and myself) in the future.